reuse view model for both screens
